Cleanup and fixed the strange behaviour with nested envs
A definition with nested names had unexpected behaviour:

    (devine <var>         <exp>)
    (define /foo/bar/baz2 foo/bar/baz)

-- here `<exp> = foo/bar/baz` was evaluated in the nested chain of
namespaces defined by `<var> = /foo/bar/baz2`. The deepest namespace
being `bar`, the name `foo` was found starting from `bar`:

    foo={} - bar={}
          \- foo=5

-- the closes `foo` is the `foo=5` in the `bar.outer`, it was picked up.

Now, both `<var>` and `<exp>` are evaluated in the original namespace of
`define` -- all variable names are found in there. Then if the result of
`<exp>` is a not anonymous Env (= outer != None), it is re-attached to
the deepest nested Env of `<var>`.

It should be more readable this way.

	modified:   nsp_lis.py
	modified:   test_namespaces_nested.py