The working standard function in metalisp tests!
Made eval forms coherent with other ones.
Though the implementation is temporary, awekward, but simple.

The Python callables are expected to be purely functional,
with no reference to the namespaces.
They could be used for the 1-step macro calls.
But now their arguments are fully evaled recursively for them, as in
usual eval.

The eval forms are run like custom ones,
but instead of eval-ing the _proc lists
they are Python callables, called with a reference to the namespace.

Eval can be an object, inheriting from the Namespace.

Exported define and eval in the global namespace.

	изменено:      sym_lis2.py
	изменено:      tests/test_metalisp.py