Made the custom explicit_eval!
The proble is that _dyn does not nest.
So it cannot be evoked in nested calls, such as recursion.
Therefore, all recursion data is passed through arguments.

Another proble:
the nested eval that does everything in only 1 namespace,
when the manipulation of several namespaces is needed.
For now I made `eval2` for this, but probably there is a better
solution.

	modified:   sym_lis2.py
	modified:   tests/test_metalisp.py
	modified:   tests/test_metalisp_basic_func.py