switching to sym_lis3
Back to the usual lexical rules for eval, but with dyn_env and root_env
in the name space. Let's develop the idea of manipulating namespaces
more.

Maybe the only downside of this is lack of flexibility in user's input:
you have to explicitly input `"string"` or `(quote (expr))` into your macro,
it cannot just parse input `string` or `expr` according to its own
semantics.

But on the other hand, it is obvious and customary, and it is clear when
you generate the `"string"` with some `(generate_string)`.

	modified:   sym_lis2.py
	new file:   sym_lis3.py
	modified:   sym_repl.py
	modified:   tests/test_metalisp_basic_func.py
	modified:   tests/test_repl.py